refactor: extract protocols.c from somewm.c
Move layer shell, idle inhibit, session lock, foreign toplevel, and
XDG activation token handlers into a dedicated protocols module.

31 functions extracted across 5 protocol groups:
- Layer shell: arrangelayer, arrangelayers, createlayersurface,
  commitlayersurfacenotify, destroylayersurfacenotify, unmaplayersurfacenotify
- Idle inhibit: checkidleinhibitor, createidleinhibitor,
  destroyidleinhibitor, some_is_idle_inhibited
- Session lock: locksession, unlocksession, createlocksurface,
  destroylocksurface, destroylock, destroysessionlock,
  some_activate/deactivate_lua_lock, some_promote_lock_cover,
  some_clear_pre_lock_client, some_is_ext_session_locked
- Foreign toplevel: all 5 request handlers
- Activation tokens: create, cleanup, cancel_all, timeout, urgent

somewm.c: 6799 -> 6053 lines (-746)
protocols.c: 828 lines (new)
